# simpler + faster curves to segments, leak fix