--- a/experiments/benchmark_svg_cairo.nim
+++ b/experiments/benchmark_svg_cairo.nim
@@ -1,6 +1,6 @@
 import benchy, svg_cairo
 
-let data = readFile("tests/images/svg/Ghostscript_Tiger.svg")
+let data = readFile("tests/fileformats/svg/Ghostscript_Tiger.svg")
 
 timeIt "svg decode":
-  keep decodeSvg(data)
+  discard decodeSvg(data)