--- a/tests/benchmark_svg.nim
+++ b/tests/benchmark_svg.nim
@@ -3,4 +3,4 @@ import benchy, pixie/fileformats/svg
 let data = readFile("tests/fileformats/svg/Ghostscript_Tiger.svg")
 
 timeIt "svg decode":
-  keep decodeSvg(data)
+  discard decodeSvg(data)