before:
`svg decode ........................ 16.072 ms     16.274 ms    ±0.186   x306`

after:
`svg decode ........................ 15.079 ms     15.133 ms    ±0.041   x330`

cairo tiger with color bgra -> rgba swap = 16.5ms (pixie wins)
cairo tiger no color swap = 12.5ms (still losing)

E must be float64 to pass leak fuzzing